Avoid skip_duplicates=True if unnecessary for correct code
As far as I can tell, duplicate entries in the code of question shouldn't
occur. If they do, it points to an issue somewhere. In this case, a warning
or error should be displayed in order to identify the issue.

In general, don't hide warnings or error unless they are expected to occur and
don't indicate that something's wrong with the code. Disabling such errors or
warnings warrants a comment in the code.