Speed up Stimulus.make() by replacing insert1 by batch insert
The change uses list comprehension to create a single data structure that
contains all trials, and then inserts it into the database with a single
self.insert(.) command. This is faster than inserting the data for each trial
separately.

Generally, we should always try to use batch insert for large'ish datasets
whenever possible.