Speed up inserting neural data into Neuralactivity table
The previous implementation of Neuralactivity.make(.) was extremely slow as it
re-loaded the whole dataset for each trial_id separately. This is a limitation
of having Neuralactivity be a dj.Imported table with Stimulus forming the
foreign part of the primary key.

I didn't see a good way to speed up the process while perserving the
table type. Instead, I here moved data loading to one single spot, namely
Stimulus.make(.) and made it add data to Neuralactivity (which is now a
dj.Manual table). I think this is consistent, as it is the only place that
is only called once for each new session.

Inserting neural data as a single batch led to database connectivity issues.
That was probably caused by the single insert command taking too long. I worked
around this by splitting the data into multiple, large'ish batches that are
inserted consecutively.

The batch size itself is controlled by the added global constant
MAX_BATCH_INSERT_SIZE.

Note that I added a comment into the code, justifying why the batch is split
into smaller chunks to explain to any reader of the code why this is done
(good practice in general for code that isn't immediately obvious).

Overall, this cut the time to insert the data of one session to ~2min on my
5-year old laptop.