Specificy data folder as global constant
Generally, any 'magic constants' ought to be avoided in the code. One example
is the data folder that might show up multiple times in the code. I moved it to
the top as a global constant. This ensures that there is only a single place
where it needs to be changed, if such a change is necessary. It also
makes the code more transparent.